feat: доработка Item и работа с Booking

+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -1,6 +1,8 @@
 package ru.practicum.shareit.user.service;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Isolation;
+import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.shareit.exception.DuplicateEmailException;
 import ru.practicum.shareit.exception.DuplicateUserException;
 import ru.practicum.shareit.exception.UserNotFoundException;
@@ -8,11 +10,11 @@ import ru.practicum.shareit.user.dto.UpdatedUserDto;
 import ru.practicum.shareit.user.model.UserEntity;
 import ru.practicum.shareit.user.repository.UserRepository;
 
+import java.sql.SQLException;
 import java.util.List;
 
-
 @Service
-public class UserServiceImpl implements ru.practicum.shareit.user.service.UserService {
+public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
 
@@ -20,22 +22,30 @@ public class UserServiceImpl implements ru.practicum.shareit.user.service.UserSe
         this.userRepository = repository;
     }
 
+    @Override
+    @Transactional(readOnly = true)
     public List<UserEntity> getAllUsers() {
         return userRepository.findAll();
     }
 
+    @Override
+    @Transactional(readOnly = true)
     public UserEntity getUserById(Long userId) {
         return userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException("Пользователь с данным ID не найден!"));
     }
 
+    @Override
+    @Transactional(isolation = Isolation.READ_COMMITTED)
     public UserEntity createUser(UserEntity user) {
-        if (!userRepository.existsByEmail(user.getEmail())) {
+        try {
             return userRepository.save(user);
-        } else {
+        } catch (Exception e) {
             throw new DuplicateUserException("Данный пользователь уже зарегистрирован!");
         }
     }
 
+    @Override
+    @Transactional(isolation = Isolation.READ_COMMITTED)
     public UserEntity updateUser(Long userId, UpdatedUserDto updatedUserDto) {
         UserEntity user = userRepository.findById(userId).orElseThrow(() ->
                 new UserNotFoundException("Обновляемый пользователь не найден!"));
@@ -54,7 +64,13 @@ public class UserServiceImpl implements ru.practicum.shareit.user.service.UserSe
         return userRepository.save(user);
     }
 
+    @Override
+    @Transactional(isolation = Isolation.READ_COMMITTED)
     public void deleteUser(Long userId) {
-        userRepository.deleteById(userId);
+        if (userRepository.existsById(userId)) {
+            userRepository.deleteById(userId);
+        } else {
+            throw new UserNotFoundException("Пользователь не зарегистрирован!");
+        }
     }
 }
